fix(x64,client): clear all client x64 compile errors
- EterLib/Pool.h: operator new first param unsigned int -> size_t (x64
  requires size_t; cleared ~148 C2821/C2333 across EterGrnLib/EterLib/
  SpeedTreeLib).
- EterGrnLib/MilesLib/SpeedTreeLib StdAfx.h: _WIN64-empty nanomites (B9).
- PRTerrainLib/TerrainType.h: PR_FLOAT_TO_INTASM FISTP -> lrintf on x64
  (no inline asm; round-to-nearest matches FISTP) (B8).
- EterLib/MSWindow.cpp: GetWindowLong/SetWindowLong + GWL_USERDATA ->
  GetWindowLongPtr/SetWindowLongPtr + GWLP_USERDATA, (LONG_PTR)this (B5).
- ScriptLib/PythonLauncher.cpp: PyObject_AsCharBuffer len int -> Py_ssize_t.
- UserInterface/StdAfx.h: guard _USE_32BIT_TIME_T to !_WIN64.

Client now compiles as x64 with 0 errors; ~2290 warnings remain (the
pointer/size/Python-marshaling pass).

+++ Lead-Client-Source/EterLib/Pool.h
@@ -245,7 +245,7 @@ class CPooledObject
 		{
 		}
 
-        void * operator new(unsigned int /*mem_size*/)
+        void * operator new(size_t /*mem_size*/)
         {
             return ms_kPool.Alloc();
         }
@@ -436,7 +436,7 @@ class CPooledObject
 		{
 		}
 
-        void * operator new(unsigned int mem_size)
+        void * operator new(size_t mem_size)
         {
             return ms_DynamicSizePool.Alloc();
         }
